backend/requirements.txt
fastapi
uvicorn
motor
pymongo
bcrypt
passlib
python-multipart
httpx
anthropic
python-dotenv
PyJWT
email-validator
